Add experimental index sources.
These are essentially scraped from cppreference and then made into valid
source files.

The idea being that in a project, these files (or in memory versions of
them) can be loaded and indexed, then by checking the top level header
for some decl it is possible to augment the code completion suggestions
with code from headers that are available in the std library but not
necessairly included in the active TU.

It would also be possible then, assuming that the user selects one of
the completions, to automatically insert the appropriate header.

Doc comments for the headers should be provided - the contents of the
files themselves will already be documented by the library authors.

In future allow users to add their own libraries e.g. posix, linux, etc.

Standard library indexes will be treated specially.